move components from admin to global
