+++ resources/assets/js/admin/components/taxonomies/TaxonomyTermEditorCurrentTerm.vue
@@ -22,7 +22,7 @@
 <script>
 import {mapGetters} from 'vuex';
 
-import WnlTaxonomyTermWithAncestors from 'js/admin/components/taxonomies/TaxonomyTermWithAncestors';
+import WnlTaxonomyTermWithAncestors from 'js/components/global/taxonomies/TaxonomyTermWithAncestors';
 
 export default {
 	components: {
